Rename utils index to misc
