+++ packages/util/index.ts
@@ -6,7 +6,7 @@ export * from './src/config';
 export * from './src/database';
 export * from './src/job-queue';
 export * from './src/constants';
-export * from './src/index';
+export * from './src/misc';
 export * from './src/fill';
 export * from './src/events';
 export * from './src/types';
