+++ packages/util/src/misc.ts
